requirements.txt
# Core dependencies (required)
streamlit>=1.28.0
requests>=2.31.0
numpy>=1.24.0
Pillow>=9.5.0
python-dateutil>=2.8.2

# Optional dependencies for enhanced features
# Uncomment to install:

# AI and ML
# google-generativeai>=0.3.0
# torch>=2.0.0
# transformers>=4.30.0
# accelerate>=0.20.0
# bitsandbytes>=0.41.0
# llama-cpp-python>=0.2.0

# Audio processing
# pyttsx3>=2.90
# pydub>=0.25.1
# openai-whisper>=20231117

# UI enhancements
# streamlit-drawable-canvas>=0.9.0

# Development and testing
# pytest>=7.0.0
# pytest-cov>=4.0.0